support file reading in single threaded context
